app/requirements.txt
Flask==1.1.2
requests==2.24.0
ptvsd==4.3.2 # Required for debugging.
PyYAML==5.3.1
google-api-python-client==1.12.5
cryptography==2.8
